HH-118342 move jersey and servlet builders outside of NabApplicationBuilder
